requirements.txt
python-telegram-bot==20.7
requests==2.31.0
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.6
asyncio==3.4.3
python-dotenv==1.0.0
schedule==1.2.0
pymongo==4.5.0